Move ui into new ui package
