+++ app/src/main/java/com/ncnf/feed/ui/EventAdapter.java
@@ -1,4 +1,4 @@
-package com.ncnf.feed;
+package com.ncnf.feed.ui;
 
 import android.view.LayoutInflater;
 import android.view.View;
